ENH: Can now vertical markers locked to data

+++ applications/mne_analyze/libs/anShared/Model/fiffrawviewmodel.h
@@ -361,7 +361,7 @@ public:
      */
     inline qint32 numVLines() const;
 
-    void newTimeMark(const int& xpos);
+    void newTimeMark(const float& xpos);
 
     float getTimeMarks(int index) const;
